requirements.txt
# ====================================
# Multi-City GIS Portfolio Analyzer
# Requirements File
# ====================================
# 
# Note: Flexible versioning used to ensure compatibility
# with pre-built wheels on all platforms (especially Windows)
# ====================================

# Core Framework
streamlit>=1.29.0
streamlit-folium>=0.15.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
geoalchemy2>=0.14.0
alembic>=1.13.0

# Geospatial Libraries (flexible versions for Windows compatibility)
geopandas>=0.14.0
folium>=0.15.0
shapely>=2.0.0
pyproj>=3.0.0
Fiona>=1.9.0
rtree>=1.0.0

# Data Processing
pandas>=2.1.0
openpyxl>=3.1.0
xlrd>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
colorama>=0.4.6
python-multipart>=0.0.6

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# File Handling
zipfile36>=0.1.3

# Validation
pydantic>=2.5.0

# Progress Bars
tqdm>=4.66.0

# Development Tools (Optional - uncomment for development)
# pytest==7.4.3
# black==23.12.1
# flake8==6.1.0
# mypy==1.7.1
# pytest-cov==4.1.0

# Documentation (Optional)
# mkdocs==1.5.3
# mkdocs-material==9.5.2

